BAU: Refactor - move CIMIT code into cimit namespace
